Phase_II/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.13.1
pydantic[email]>=2.10.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
psutil==5.9.8
cloudinary==1.44.1
Pillow==10.2.0
slowapi==0.1.9

# Email dependencies
resend==0.8.0
aiosmtplib==3.0.1
email-validator==2.1.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-env==1.1.3
httpx==0.25.2

